+++ pkgs/test_api/test/frontend/matcher/throws_type_test.dart
@@ -42,8 +42,9 @@ void main() {
 
   group('[throwsCyclicInitializationError]', () {
     test('passes when a CyclicInitializationError is thrown', () {
-      // ignore: deprecated_member_use
-      expect(() => throw CyclicInitializationError(''),
+      expect(
+          () => _CyclicInitializationFailure().x,
+          // ignore: deprecated_member_use
           throwsCyclicInitializationError);
     });
 
@@ -53,9 +54,7 @@ void main() {
       });
 
       expectTestFailed(
-          liveTest,
-          startsWith(
-              "Expected: throws <Instance of 'CyclicInitializationError'>"));
+          liveTest, startsWith("Expected: throws <Instance of 'Error'>"));
     });
   });
 
@@ -103,26 +102,6 @@ void main() {
 
       expectTestFailed(liveTest,
           startsWith("Expected: throws <Instance of 'NoSuchMethodError'>"));
-    });
-  });
-
-  group('[throwsNullThrownError]', () {
-    test('passes when a NullThrownError is thrown', () {
-      // Throwing null is no longer allowed with NNBD, but we do want to allow
-      // it from legacy code and should be able to catch those errors.
-      // ignore: deprecated_member_use
-      expect(() => throw NullThrownError(), throwsNullThrownError);
-    });
-
-    test('fails when a non-NullThrownError is thrown', () async {
-      var liveTest = await runTestBody(() {
-        expect(() => throw Exception(), throwsNullThrownError);
-      });
-
-      expectTestFailed(
-          liveTest,
-          // ignore: deprecated_member_use
-          startsWith("Expected: throws <Instance of '$NullThrownError'>"));
     });
   });
 
@@ -185,4 +164,9 @@ void main() {
           startsWith("Expected: throws <Instance of 'UnsupportedError'>"));
     });
   });
+}
+
+class _CyclicInitializationFailure {
+  late int x = y;
+  late int y = x;
 }